docker/ai-core/requirements.txt
# ------------------------------
# Core ML & Data Processing
# ------------------------------
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
scipy==1.13.0
joblib==1.4.2

# ------------------------------
# Database & Data IO
# ------------------------------
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9

# ------------------------------
# API / Orchestration Layer
# ------------------------------
fastapi==0.111.0
uvicorn==0.30.1
requests==2.32.3

# ------------------------------
# Model Management / Serialization
# ------------------------------
pydantic==2.7.1
pyyaml==6.0.2

# ------------------------------
# Automation & Scheduling
# ------------------------------
apscheduler==3.10.4

# ------------------------------
# Optional Multi-Modal (Future Expansion)
# ------------------------------
#transformers==4.43.2
#torch==2.2.0
#sentence-transformers==2.6.0
#opencv-python==4.9.0.80
#Pillow==10.2.0

# ------------------------------
# Utilities
# ------------------------------
loguru==0.7.2
python-dotenv==1.0.1
